fix: 쿼리 분리
